backend/requirements.txt
# FastAPI
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File uploads
python-multipart==0.0.19

# Settings
pydantic-settings==2.6.1

# Validation
email-validator==2.1.0

# Utils
python-dotenv==1.0.1

